Move tests into framework-specific packages
